fix: initialize the standalone Desktop preview backend
